fix aids ridden spaghetti structure
