+++ packages/synapse-interface/components/Pools/pool/NoPoolBody.tsx
