+++ packages/synapse-interface/components/Pools/pool/PoolInfoSection/CurrencyReservesCard.tsx
