+++ packages/synapse-interface/components/Pools/pool/PoolInfoSection/InfoSection.tsx
