+++ packages/synapse-interface/components/Pools/pool/PoolInfoSection/InfoSectionCard.tsx
