+++ packages/synapse-interface/components/Pools/pool/components/PoolTitle.tsx
