+++ packages/synapse-interface/components/Pools/pool/components/ReceivedTokenSection.tsx
