+++ packages/synapse-interface/components/landing/sections/HeroSection.tsx
