+++ packages/synapse-interface/components/landing/sections/HowItWorksSection.tsx
