+++ packages/synapse-interface/components/landing/sections/IntegrationSection.tsx
