+++ packages/synapse-interface/components/landing/sections/ResourcesSection.tsx
