+++ packages/synapse-interface/components/landing/sections/SecuritySection.tsx
