+++ packages/synapse-interface/components/stake/StakeCardTitle.tsx
