+++ packages/synapse-interface/pages/pool/[poolId].tsx
@@ -6,8 +6,8 @@ import { useRouter } from 'next/router'
 import StandardPageContainer from '@layouts/StandardPageContainer'
 import { LandingPageWrapper } from '@layouts/LandingPageWrapper'
 import { DEFAULT_FROM_CHAIN } from '@/constants/swap'
-import PoolBody from './PoolBody'
-import NoPoolBody from './NoPoolBody'
+import PoolBody from '@/components/Pools/pool/PoolBody'
+import NoPoolBody from '@/components/Pools/pool/NoPoolBody'
 import { fetchPoolData, resetPoolData } from '@/slices/poolDataSlice'
 import { usePoolDataState } from '@/slices/pool/hooks'
 import { resetPoolDeposit } from '@/slices/poolDepositSlice'
