+++ packages/synapse-interface/pages/stake/[routerIndex].tsx
@@ -2,7 +2,7 @@ import { useAccount, useNetwork } from 'wagmi'
 import { useRouter } from 'next/router'
 import StandardPageContainer from '@layouts/StandardPageContainer'
 import { LandingPageWrapper } from '@layouts/LandingPageWrapper'
-import StakeCard from './StakeCard'
+import StakeCard from '@/components/stake/StakeCard'
 import { POOL_BY_ROUTER_INDEX } from '@/constants/tokens'
 import Link from 'next/link'
 import { POOL_PATH } from '@/constants/urls'
